Splits out util functions
